Store all configuration in structures rather than constants. Change single bits instead of bytes where needed.
